Applying all changes to Cmgui revisions 9240 through 9359 to branch to_libzinc_and_beyond.  https://tracker.physiomeproject.org/show_bug.cgi?id=3359
git-svn-id: https://svn.physiomeproject.org/svn/cmiss/cmgui/branches/to_libzinc_and_beyond@9363 3c1deb5b-d424-0410-962d-aba41a686d42

+++ core/source/computed_field/field_cache.hpp
@@ -267,13 +267,13 @@ public:
 	// NOT THREAD SAFE
 	void setValueCache(int cacheIndex, FieldValueCache* valueCache)
 	{
-		if (cacheIndex < (int)valueCaches.size())
+		if (cacheIndex < static_cast<int>(valueCaches.size()))
 		{
 			delete valueCaches[cacheIndex];
 		}
 		else
 		{
-			for (int i = valueCaches.size(); i <= cacheIndex; ++i)
+			for (int i = static_cast<int>(valueCaches.size()); i <= cacheIndex; ++i)
 				valueCaches.push_back(0);
 		}
 		valueCaches[cacheIndex] = valueCache;
